# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: M-Sklad Restaurant Inventory & ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /inventory-audits:
    get:
      operationId: listInventoryAudits
      tags: [inventory-audits]
      summary: List inventory audits
      responses:
        "200":
          description: Audits list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryAudit"
    post:
      operationId: createInventoryAudit
      tags: [inventory-audits]
      summary: Start a new inventory audit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryAuditBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryAudit"
  /inventory-audits/{id}:
    get:
      operationId: getInventoryAudit
      tags: [inventory-audits]
      summary: Get an inventory audit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryAudit"
    patch:
      operationId: updateInventoryAudit
      tags: [inventory-audits]
      summary: Update audit items (submit counts)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryAuditBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryAudit"
  /inventory-audits/{id}/submit:
    post:
      operationId: submitInventoryAudit
      tags: [inventory-audits]
      summary: Submit (finalize) an inventory audit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryAudit"
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - createdAt
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
      required:
        - name
        - slug
    UpdateCategoryBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
...
    InventoryAudit:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [draft, submitted]
        items:
          type: array
          items:
            $ref: "#/components/schemas/AuditItem"
        createdAt:
          type: string
          format: date-time
        submittedAt:
          type: ["string", "null"]
      required:
        - id
        - title
        - status
        - items
        - createdAt
    AuditItem:
      type: object
      properties:
        id:
          type: integer
        auditId:
          type: integer
        itemId:
          type: integer
        itemName:
          type: ["string", "null"]
        systemStock:
          type: number
        actualStock:
          type: ["number", "null"]
        discrepancy:
          type: ["number", "null"]
      required:
        - id
        - auditId
        - itemId
        - systemStock
    CreateInventoryAuditBody:
      type: object
      properties:
        title:
          type: string
        categoryId:
          type: ["integer", "null"]
      required:
        - title
    UpdateInventoryAuditBody:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              itemId:
                type: integer
              actualStock:
                type: number
            required:
              - itemId
              - actualStock
      required:
        - items
...